requirements.txt
django
djangorestframework
pygments
django-import-export
django-rest-swagger
django-extensions
pylint
ipykernel
pandas
unidecode